backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.2,<2.7
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0

# Vector search & ML

faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy>=1.25.0,<2.0
scikit-learn==1.3.2
huggingface-hub<0.20.0


# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
aiosqlite==0.19.0  # Async SQLite

# Security & Rate limiting
python-multipart==0.0.6
slowapi==0.1.9
pydantic-extra-types==2.3.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
PyJWT>=2.8.0,<3.0


# Utilities
aiofiles==23.2.1
httpx==0.25.1
pytz==2023.3
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Logging & monitoring
python-json-logger==2.0.7
structlog==23.2.0

# Data validation
validators==0.22.0
phonenumbers==8.13.0

torch==2.1.0
